--- src/components/ButtonBurger/ButtonBurgerCC/interface.ts
@@ -1,6 +1,0 @@
-import { IMobileMenuState } from '../../../store/slices/mobileMenuSlice';
-
-export interface IButtonBurgerCC {
-  mobileMenu: IMobileMenuState;
-  toggleMobileMenu: any;
-}
