+++ src/components/buttonTheme/ButtonThemeCC/interface.ts
@@ -1,0 +1,6 @@
+import { IThemeState } from '../../../store/slices/themeSlice';
+
+export interface IButtonThemeCC {
+  theme: IThemeState;
+  switchTheme: any;
+}
